Fix a minor error in the Mongolian Cyrillic IME
If we use a genuine Mongolian Cyrillic keyboard, when we press on the button "'", the outcome should be "п".
But in this IME, the actual outcome is "'". However, when we press on the button "\" by using the IME, the outcome is "п".
(It should be "\".) I'm doing this to fix the errors.