Fix broken timer initialization.
init() calls position() which calls focus(), which sets a timer.
Because the timer was set to null right after the init() call,
it was not properly cleared out. This caused a UX issue on first
load, where the IME selector would disappear right under the
mouse in spite of the recent fix to the mouseenter behavior.

It's questionable whether we want to set that timer at all on
init, but since it's now properly stopped, it's harmless.