package.json: Bump all devDependencies to latest, pull in libraries
Bump .jshintrc to a more current version (though lots of this is using
outdated and soon-to-be-removed options); some of the already-removed
options have been dropped from it for now, but will require revisiting.

No need to load jQuery and Qunit over the Web when we can trivially
just add them as devDependencies (like we do elsewhere).