Add Mon Simplified Anonta keyboard
Requested in translatewiki:
https://translatewiki.net/wiki/Thread:Support/Want_to_ask_two_quick_questions

It's based on the Mon Anonta layout distributed
by Keyman:
https://help.keyman.com/keyboard/mon_anonta/1.0.1/mon_anonta.php

The main change is that the use of VOWEL SIGN E on the 'a'
key is simplified, so it's typed after the vowel and not before it.
Getting it to be typed before the letter is a bit too complicated
for jquery.ime.