jv-kongres keyboard (#729)
* jv-kongres keyboard

This is a rather simple transliteration keyboard that resulted from the Kongres Aksara Jawa I 2021. Users are required to memorize a lot of keys, and the typing speed is subpar from Nulisa transliteration method (jv-transliteration).

* Update jv-kongres.js

* Update jquery.ime.test.fixtures.js

* adding test cases

* Update jquery.ime.test.fixtures.js

escape characters

* Update jquery.ime.test.fixtures.js

try another test

* Squashing bug

What an elusive bug! There should be a documentation that these special characters need to be escaped!

* Update jquery.ime.test.fixtures.js

* Update rules/jv/jv-kongres.js

Thanks @labdajiwa

Co-authored-by: Labdajiwa <labdajiwa@protonmail.com>

* Update .gitignore

* Update .gitignore

* add jv-kongres to test and definitions

* Update .gitignore

* Update .gitignore

* Update jquery.ime.test.fixtures.js

---------

Co-authored-by: Labdajiwa <labdajiwa@protonmail.com>
Co-authored-by: SrishAkaTux <ssethi@wikimedia.org>
Co-authored-by: SrishAkaTux <srishakatux@gmail.com>